[compiler] Remove section on healthcheck (#7239)
* [compiler] Remove section on healthcheck

This package will be deprecated soon. It makes adopting the compiler
confusing since it can spread the misconception that you need to have
all your components successfully compiled before you can use the
compiler.

For now let's remove this from our docs and deprecate the package on
npm. We can always choose to repurpose this in the future.

* Clarify that 100% compilation is not necessary